.engine => .graphEngine to maybe stop weird bug???
After working on TW5-Graph for a while, OTHER Tiddlywikis will get
a bug where they'll think *.engine is undefined. I have no idea
how that's happening. I've looked at the code, and it should be=
impossible. I can only assume it's some firefox data corruption.
So I'm changing the name of this variable to see if it stops it.