Bumping the attached vis-network to V10.0.0
It's just for stupid technical reasons that the version needs to
bump this much. The old visjs library was at 9.something, and then
I put this at 9.something before I added its own adapter stuff
which is the only code that's going to evolve anymore. If I drop
lower, it may interfere with some people being able to upgrade
automatically.